Consolidate parameter setters
