--- modules/gltools/src/state-tracker/state-tracking/index.js
@@ -1,6 +1,0 @@
-export {
-  default as trackContextState,
-  pushContextState,
-  popContextState
-} from './track-context-state';
-export {withParameters} from './with-parameters';
