+++ modules/gltools/test/state-tracker/state-tracking/context-state.spec.js
@@ -4,7 +4,7 @@ import {getParameters, setParameters, resetParameters, withParameters} from '@lu
 
 import {getKey, Framebuffer} from '@luma.gl/webgl';
 
-import {GL_PARAMETER_DEFAULTS as GL_PARAMETERS} from '@luma.gl/gltools/state-tracker/unified-parameter-api/webgl-parameter-tables';
+import {GL_PARAMETER_DEFAULTS as GL_PARAMETERS} from '@luma.gl/gltools/state-tracker/webgl-parameter-tables';
 
 import {ENUM_STYLE_SETTINGS_SET1} from '../data/sample-enum-settings';
 import {FUNCTION_STYLE_SETTINGS_SET1} from '../data/sample-function-settings';
